requirements.txt
factory-boy>=2.12.0
Flask>=1.0.2
Flask-Cors>=3.0.7
Flask-SQLAlchemy>=2.4.0
Flask-WTF>=0.14

